Merged Session into User
Basically added functionality of session into user, and tweaked it up a
lot for it to work. You sign up with userId, email, password, and class.
Login with email and password. Logging in with previous stuff may not
work though cause of different cell types then data types being inputed.