Subclass specific methods implemented
Basically fixed how to get subclass specific attributes, where its not
null if its that class but null in other classes, and added the specific
methods to user, but left empty, and overridden in specific sub classes.
Also added a temporary way to verify someone. Also added test cases for
it and it passes all the tests.